+++ packages/admin/admin-date-time/src/index.ts
@@ -3,4 +3,5 @@ export { DateRange, DateRangePicker, DateRangePickerProps } from "./DateRangePic
 export { FinalFormDatePicker, FinalFormDatePickerProps } from "./FinalFormDatePicker";
 export { FinalFormDateRangePicker, FinalFormDateRangePickerProps } from "./FinalFormDateRangePicker";
 export { FinalFormTimePicker } from "./FinalFormTimePicker";
+export { DateFnsLocaleContext, DateFnsLocaleProvider, useDateFnsLocale } from "./helpers/DateFnsLocaleProvider";
 export { TimePicker, TimePickerProps } from "./TimePicker";